+++ src/main/java/it/gov/pagopa/pu/pagopapayments/connector/pagopa/taxonomy/client/PagoPaTaxonomyApiClient.java
@@ -3,7 +3,7 @@ package it.gov.pagopa.pu.pagopapayments.connector.pagopa.taxonomy.client;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import it.gov.pagopa.pu.pagopapayments.config.RestTemplateConfig;
+import it.gov.pagopa.pu.pagopapayments.config.rest.RestTemplateConfig;
 import it.gov.pagopa.pu.pagopapayments.connector.pagopa.taxonomy.dto.PaTaxonomyDTO;
 import it.gov.pagopa.pu.pagopapayments.exception.ApplicationException;
 import lombok.extern.slf4j.Slf4j;
